File was missing from initial commit
This crontab file runs once a day.

A couple of things to mention about some of the fields in the output json file:

* maintenance_mode was never really used, but if you want to switch the app to maintenance_mode then this is the flag to do it. It's been a long time since it was tested but it should set some of the app into an offline mode while allowing the user to access some features still.
* ssl isn't used because it's hard coded in the ChastiKey app code as 1 (true).

The rest of the fields are mainly used for bot stats to show on the lock creation screen when selecting a bot. I chose to run the crontab daily because the bot stats shown didn't need to have up to the minute lock counts and ratings.